AP_IOMCU: fixed delay in DSM bind
